Refactor cli/gui architecture
- Move cli related files to a cli folder
- Move parameter validation from parameter_manager.py to parameter_check.py
- Move exporter from ui folder to wider scope
- Move time estimation from parameter_manager.py to time_estimation.py, and extend for displacement and analysis.
- Fix cli crash by giving progress_callback without argument
- Add time indication in cli
- Take care of documentation
