+++ app/gui/parameter_manager.py
@@ -9,6 +9,9 @@ from PySide6.QtWidgets import QMessageBox
 from matplotlib.colors import to_hex
 
 
+from app.time_estimation import TimeEstimation
+
+
 class ParameterManagerMixin:
     """Mixin for MainWindow to handle parameter gathering, validation, and equivalency checks."""
 
@@ -137,123 +140,6 @@ class ParameterManagerMixin:
 
         return params
 
-    def _get_time_estimation_indicators(self, params: dict) -> tuple[str, str]:
-        """
-        Calculate optimization time estimation indicators.
-
-        Parameters
-        ----------
-        params : dict
-            Current parameters dictionary.
-
-        Returns
-        -------
-        tuple[str, str]
-            (color, tooltip_text) - Color hex code and tooltip description
-            indicating estimated runtime.
-        """
-        nelx: int
-        nely: int
-        nelz: int
-        nelx, nely, nelz = self.last_params.get("Dimensions", {}).get(
-            "nelxyz", [1, 1, 1]
-        )
-        is_3d: bool = nelz > 0
-        dim = 3 if is_3d else 2
-
-        # Get sizeMatrix (get number of fixed dofs for it)
-        supports: dict = self.last_params.get("Supports", {})
-        fixed_dofs: int = 0
-        if supports:
-            sx: list = supports.get("sx", [])
-            sy: list = supports.get("sy", [])
-            sz: list = supports.get("sz", [])
-            sr: list = supports.get("sr", [0] * len(sx))
-            sdim: list = supports.get("sdim", [])
-            active_sup: list = [i for i, val in enumerate(sdim) if val != "-"]
-
-            for i in active_sup:
-                fixed_dofs += 1
-
-                if i < len(sr) and sr[i] > 0:
-                    radius = sr[i]
-                    x_range = range(
-                        max(0, int(sx[i] - radius)),
-                        min(nelx + 1, int(sx[i] + radius + 1)),
-                    )
-                    y_range = range(
-                        max(0, int(sy[i] - radius)),
-                        min(nely + 1, int(sy[i] + radius + 1)),
-                    )
-                    z_range = (
-                        range(
-                            max(0, int(sz[i] - radius)),
-                            min(nelz + 1, int(sz[i] + radius + 1)),
-                        )
-                        if is_3d
-                        else range(1)
-                    )
-
-                    for z in z_range:
-                        for x in x_range:
-                            for y in y_range:
-                                dist_sq = (
-                                    (x - sx[i]) ** 2
-                                    + (y - sy[i]) ** 2
-                                    + ((z - sz[i]) ** 2 if is_3d else 0)
-                                )
-                                if dist_sq <= radius**2:
-                                    fixed_dofs += 1
-                coef = 0
-                if "X" in sdim[i]:
-                    coef += 1
-                if "Y" in sdim[i]:
-                    coef += 1
-                if is_3d and "Z" in sdim[i]:
-                    coef += 1
-                fixed_dofs *= coef
-
-        sizeMatrix: int = (nelx + 1) * (nely + 1) * (nelz + 1 if is_3d else 1) * dim
-        if sizeMatrix < 0:
-            sizeMatrix = 0
-
-        # Get nbIteration
-        forces: dict = self.last_params.get("Forces", {})
-        nbIteration: int = len([d for d in forces.get("fidir", []) if d != "-"]) + len(
-            [d for d in forces.get("fodir", []) if d != "-"]
-        )
-
-        # Get nbSolves
-        nbSolves: int = self.last_params.get("Optimizer", {}).get("n_it", 50)
-
-        def time_estimated(nbIter: int, nbSolves: int, sMatrix: int) -> float:
-            # 1.1 to add 10% of the time to consider filtering, ...
-            # NbIteration is the number of force (both input and output)
-            # NbIteration+1 to consider preparation time (which is roughly as long as an iteration)
-            # SizeMatrix is the number of degree of freedom: (nelx+1)(nely+1)(nelz+1)*2 - fixed degree of freedom (see supports)
-            # SizeMatrix^1.5 is the solver time for sparse matrix
-            return 1.1 * (nbIter + 1) * nbSolves * (sMatrix**1.5)
-
-        est_time = time_estimated(nbIteration, nbSolves, sizeMatrix)
-
-        tooltip_text = "Start the optimization process\n\nEstimation:"
-        if est_time < 10000000:
-            color = "#00FF0D"
-            tooltip_text += " Very fast (a few seconds)"
-        elif est_time < 100000000:
-            color = "#91FF00"
-            tooltip_text += " Fast (less than a minute)"
-        elif est_time < 1000000000:
-            color = "#FBC02D"
-            tooltip_text += " Medium (a few minutes)"
-        elif est_time < 10000000000:
-            color = "#FF7300"
-            tooltip_text += " Slow (less than an hour)"
-        else:
-            color = "#FF0000"
-            tooltip_text += " Very slow (more than an hour)"
-        return color, tooltip_text
-
     def on_parameter_changed(self) -> None:
         """React when a parameter is changed."""
         # First, check if a valid result from a previous run exists.
@@ -290,7 +176,8 @@ class ParameterManagerMixin:
 
         # Play the animation and update tooltip
         self.last_params = self._gather_parameters()
-        color, tooltip_text = self._get_time_estimation_indicators(self.last_params)
+        color, label = TimeEstimation(self.last_params).optimization_indicators()
+        tooltip_text = f"Start the optimization process\n\nEstimation: {label}"
         self.footer.start_create_button_effect(color_hex=color)
         self.footer.create_button.setToolTip(tooltip_text)
 
@@ -493,241 +380,6 @@ class ParameterManagerMixin:
                     pm["color"] = ["#000000"]  # Default black color
                 if "init_type" not in pm:
                     pm["init_type"] = 0
-
-    def _validate_parameters(self, params: dict) -> str | None:
-        """
-        Validate the full parameters dictionary before running optimization.
-
-        Performs a sequence of targeted checks (domain, regions, forces,
-        supports, materials and optimizer settings). Returns an error string
-        describing the first detected issue, or `None` when the parameters are
-        valid.
-
-        Parameters
-        ----------
-        params : dict
-            Complete parameters dictionary as produced by `_gather_parameters`.
-
-        Returns
-        -------
-        str or None
-            Error message describing the invalid parameter, or `None` when
-            validation passes.
-        """
-        nelx: int
-        nely: int
-        nelz: int
-        nelx, nely, nelz = params["Dimensions"]["nelxyz"]
-        if nelx <= 0 or nely <= 0 or nelz < 0:
-            return "Nx, Ny, Nz must be positive."
-
-        # Validate initialization from current result if chosen
-        if params.get("Materials", {}).get("init_type") == 3:
-            last_x = getattr(self, "last_successful_xPhys", None)
-            if last_x is None:
-                return "No current result available to initialize from."
-            nel: int = nelx * nely * (nelz if nelz > 0 else 1)
-            last_x_nel: int = last_x.shape[1] if last_x.ndim == 2 else last_x.size
-            if last_x_nel != nel:
-                return f"Current result grid size ({last_x_nel}) does not match the active grid dimensions ({nel})."
-
-        err = (
-            self._check_domain(params)
-            or self._check_regions(params)
-            or self._check_forces(params)
-            or self._check_supports(params)
-            or self._check_materials(params)
-            or self._check_optimizer(params)
-        )
-        return err
-
-    def _check_domain(self, params: dict) -> str | None:
-        """
-        Validate domain-related configuration values.
-        """
-        pd: dict = params.get("Dimensions", {})
-        nelx: int
-        nely: int
-        nelz: int
-        nelx, nely, nelz = pd.get("nelxyz", [0, 0, 0])
-        if nelx < 0 or nely < 0 or nelz < 0:
-            return "Dimensions must not be negative."
-        volfrac: float = pd.get("volfrac", 0.0)
-        if volfrac <= 0.0 or volfrac > 1.0:
-            return "Volume fraction must be between 0 and 1."
-
-    def _check_optimizer(self, params: dict) -> str | None:
-        """
-        Validate optimizer-related parameters.
-        """
-        po: dict = params.get("Optimizer", {})
-        if po.get("filter_radius_min", 0.0) < 0.0:
-            return "Filter radius must be non-negative."
-        eta: float = po.get("eta", 0.0)
-        if eta <= 0.0:
-            return "Optimizer eta must be positive."
-        if eta > 1.0:
-            return "Optimizer eta must be at most 1.0."
-        penal: float = po.get("penal", 0.0)
-        if penal < 1.0:
-            return "Penalization must be at least 1.0."
-        if penal > 10.0:
-            return "Penalization must be at most 10.0."
-        if po.get("max_change", 0.0) <= 0.0:
-            return "Optimizer max_change must be positive."
-        if po.get("n_it", 0) <= 0:
-            return "Optimizer iteration count must be positive."
-
-    def _check_duplicates(
-        self, indices: list, keyfunc: callable, msg: callable
-    ) -> str | None:
-        """
-        Check for duplicate entries and return error message if found.
-
-        Parameters
-        ----------
-        indices : list
-            List of indices to check.
-        keyfunc : callable
-            Function to extract comparison key from index.
-        msg : callable
-            Function to generate error message (takes two indices).
-
-        Returns
-        -------
-        str or None
-            Error message if duplicate found, None otherwise.
-        """
-        seen = {}
-        for i in indices:
-            k = keyfunc(i)
-            if k in seen:
-                return msg(seen[k], i)
-            seen[k] = i
-
-    def _check_forces(self, params: dict) -> str | None:
-        """
-        Validate that at least one input force is active.
-
-        Parameters
-        ----------
-        params : dict
-            Parameters dictionary.
-
-        Returns
-        -------
-        str or None
-            Error message if validation fails, None otherwise.
-        """
-        pf: dict = params["Forces"]
-        ps: dict = params.get("Supports", {})
-
-        if not any(d != "-" for d in pf["fidir"]):
-            return "At least one input force must be active"
-
-        if not any(d != "-" for d in pf["fodir"]) and not any(
-            d != "-" for d in ps.get("sdim", [])
-        ):
-            return "At least one output force (for compliant mechanisms) or support (for rigid mechanisms) must be active"
-
-        err: str | None = self._check_duplicates(
-            [i for i, d in enumerate(pf["fidir"]) if d != "-"],
-            lambda i: (pf["fix"][i], pf["fiy"][i], pf["fiz"][i], pf["fidir"][i]),
-            lambda a, b: f"Input forces {a + 1} and {b + 1} are identical.",
-        )
-        if err:
-            return err
-
-        return self._check_duplicates(
-            [i for i, d in enumerate(pf["fodir"]) if d != "-"],
-            lambda i: (pf["fox"][i], pf["foy"][i], pf["foz"][i], pf["fodir"][i]),
-            lambda a, b: f"Output forces {a + 1} and {b + 1} are identical.",
-        )
-
-    def _check_regions(self, params: dict) -> str | None:
-        """
-        Check for duplicate regions.
-
-        Parameters
-        ----------
-        params : dict
-            Parameters dictionary.
-
-        Returns
-        -------
-        str or None
-            Error message if duplicate found, None otherwise.
-        """
-        pr: dict = params.get("Region")
-        if not pr:
-            return
-
-        idx = [i for i, s in enumerate(pr["rshape"]) if s != "-"]
-
-        return self._check_duplicates(
-            idx,
-            lambda i: (
-                pr["rshape"][i],
-                pr["rradius"][i],
-                pr["rx"][i],
-                pr["ry"][i],
-                pr["rz"][i],
-            ),
-            lambda a, b: f"Regions {a + 1} and {b + 1} are identical.",
-        )
-
-    def _check_supports(self, params: dict) -> str | None:
-        """
-        Check for duplicate supports.
-
-        Parameters
-        ----------
-        params : dict
-            Parameters dictionary.
-
-        Returns
-        -------
-        str or None
-            Error message if duplicate found, None otherwise.
-        """
-        ps: dict = params.get("Supports")
-        if not ps:
-            return
-
-        idx: list = [i for i, s in enumerate(ps["sdim"]) if s != "-"]
-
-        return self._check_duplicates(
-            idx,
-            lambda i: (ps["sx"][i], ps["sy"][i], ps["sz"][i], ps["sdim"][i]),
-            lambda a, b: f"Supports {a + 1} and {b + 1} are identical.",
-        )
-
-    def _check_materials(self, params: dict) -> str | None:
-        """
-        Check for duplicate materials and validate percentages sum to 100.
-
-        Parameters
-        ----------
-        params : dict
-            Parameters dictionary.
-
-        Returns
-        -------
-        str or None
-            Error message if validation fails, None otherwise.
-        """
-        pm: dict = params["Materials"]
-
-        err = self._check_duplicates(
-            range(len(pm["E"])),
-            lambda i: (pm["E"][i], pm["nu"][i], pm.get("percent", [100])[i]),
-            lambda a, b: f"Materials {a + 1} and {b + 1} are identical.",
-        )
-        if err:
-            return err
-
-        if len(pm["E"]) > 1 and sum(pm["percent"]) != 100:
-            return "Material percentages don't sum up to 100%."
 
     def _update_position_ranges(self) -> None:
         """
