+++ app/gui/resource_path_finder.py
