+++ pyproject.toml
@@ -44,7 +44,7 @@ Issues = "https://github.com/ninja7V/topoptcomec/issues"
 topoptcomec = "app.__main__:main"
 
 [tool.setuptools]
-packages = ["app", "app.core", "app.ui"]
+packages = ["app", "app.core", "app.gui"]
 include-package-data = true
 
 [tool.setuptools.package-data]
